0.0.0  VerifyConfirm.c 0.0.0 morph of rayConfirm.c
This version will be used at the initiation of a set of verifications as part of successfully building a new cache.  There will be arrangements to only do this once per run.  There might be a
way to hush execution of the VerifyConfirm.exe when the cache is known to have not been changed.